prepared for cmake subproject build
A user might either use the git subproject approach or he/she might
clone umurmur and umurmur-monitor side-by-side like so:

  some_base_dir
        |
        +-> umurmur
        +-> umurmur-monitor

This allows umurmur-monitor to find the umurmur headers needed to build
the project. If for some reason none of the above choices is
acceptable, the user might specify UMURMUR_ROOT_PATH like so:

  cmake -DUMURMUR_ROOT_PATH=/path/to/source/of/umurmur <other_args>

and build umurmur-monitor separately from umurmur.